requirements.txt
requests>=2.32.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

